debug

+++ data/dataset.py
@@ -33,7 +33,7 @@ class H5Dataset(torch.utils.data.Dataset):
         breakpoint()
         Xk, idx = self.idx_to_key[idx]
         Yk = Xk.replace("X", "Y")
-        return self.h5f[Xk][idx], self.h5f[Yk][0, idx]
+        return self.h5f[Xk][idx].float(), self.h5f[Yk][0, idx]
 
 
 def test_H5Dataset():
